@ Extract gate-stub construction into scripts/tickets_dispatch_gate.py
tickets_dispatch.py stood at 508 of its 510-line ceiling, so the gate
family moves whole into its own module before it grows: the stub
frontmatter, the four cut-time sections each kind states, and the two
command halves that read the run and write the family all-or-none.
No behavior changes here.

The sealed HEAD probe becomes an argument the facade supplies. One
revision has to reach every stub a family writes, and naming the seam
where the family is assembled keeps that property stated rather than
incidental.

The module is imported by name (non-package branch) and relatively
(package branch) so the pinned import census in the ticket family is
untouched, the same convention tickets_bound and tickets_reissue use.

Ticket: 00-root.07 (run 20260824T222500Z-emission-law)
@
